requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
future==0.17.1
idna==2.7
pickleDB==0.8.0
pycparser==2.19
python-telegram-bot==11.1.0
requests==2.20.1
simplejson==3.16.0
six==1.11.0
upyun==2.5.2
urllib3==1.24.1
